no rewards if energy decreased
